app/server/pyproject.toml
[tool.poetry]
name = "stock-trends-server"
version = "0.1.0"
description = "Web Server for Strock Trends Strategy Project"
authors = ["Christian Kozalla <christian.kozalla@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
databases = "^0.8.0"
asyncpg = "^0.28.0"
psycopg2-binary = "^2.9.7"
aiosqlite = "^0.19.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
